+++ resources/views/admin/types/index.blade.php
@@ -8,9 +8,9 @@
 
 @section('content')
 @include('larena-content::admin.partials.editorial-steps')
-<section class="larena-admin-card">
-  <div class="larena-admin-table-wrap"><table class="larena-admin-table"><thead><tr><th>{{ __('larena-content::admin.fields.key') }}</th><th>{{ __('larena-content::admin.fields.label') }}</th><th>{{ __('larena-content::admin.fields.version') }}</th><th>{{ __('larena-content::admin.fields.fields') }}</th></tr></thead><tbody>
-  @forelse($types as $type)<tr><td><code>{{ $type['key'] }}</code></td><td>{{ $type['label'] }}</td><td>{{ $type['version'] }}</td><td>{{ $type['fields'] }}</td></tr>@empty<tr><td colspan="4">{{ __('larena-content::admin.types.empty') }}</td></tr>@endforelse
+<section class="larena-panel">
+  <div class="larena-table-scroll" tabindex="0"><table class="larena-table larena-table-stack"><thead><tr><th>{{ __('larena-content::admin.fields.key') }}</th><th>{{ __('larena-content::admin.fields.label') }}</th><th>{{ __('larena-content::admin.fields.version') }}</th><th>{{ __('larena-content::admin.fields.fields') }}</th></tr></thead><tbody>
+  @forelse($types as $type)<tr><td data-label="{{ __('larena-content::admin.fields.key') }}"><code>{{ $type['key'] }}</code></td><td data-label="{{ __('larena-content::admin.fields.label') }}">{{ $type['label'] }}</td><td data-label="{{ __('larena-content::admin.fields.version') }}">{{ $type['version'] }}</td><td data-label="{{ __('larena-content::admin.fields.fields') }}">{{ $type['fields'] }}</td></tr>@empty<tr><td colspan="4">{{ __('larena-content::admin.types.empty') }}</td></tr>@endforelse
   </tbody></table></div>
 </section>
 @endsection